clean (FileDownloader.cs): split the file into multiple partials for the sake of readability
