+++ Laerdal.McuMgr/Shared/FileDownloading/FileDownloader.Commandable.Downloading.cs
@@ -1,0 +1,6 @@
+namespace Laerdal.McuMgr.FileDownloading
+{
+    public partial class FileDownloader
+    {
+    }
+}
