+++ Laerdal.McuMgr/Shared/FileDownloading/FileDownloader.Commandable.TryCancel.cs
@@ -1,0 +1,7 @@
+namespace Laerdal.McuMgr.FileDownloading
+{
+    public partial class FileDownloader
+    {
+        public bool TryCancel(string reason = "") => _nativeFileDownloaderProxy?.TryCancel(reason) ?? false;
+    }
+}
