+++ Laerdal.McuMgr/Shared/FileDownloading/FileDownloader.Commandable.TryDisconnect.cs
@@ -1,0 +1,7 @@
+namespace Laerdal.McuMgr.FileDownloading
+{
+    public partial class FileDownloader
+    {
+        public bool TryDisconnect() => _nativeFileDownloaderProxy?.TryDisconnect() ?? false;
+    }
+}
