+++ Laerdal.McuMgr/Shared/FileDownloading/FileDownloader.Commandable.TryPause.cs
@@ -1,0 +1,7 @@
+namespace Laerdal.McuMgr.FileDownloading
+{
+    public partial class FileDownloader
+    {
+        public bool TryPause() => _nativeFileDownloaderProxy?.TryPause() ?? false;
+    }
+}
