+++ Laerdal.McuMgr/Shared/FileDownloading/FileDownloader.Commandable.TryResume.cs
@@ -1,0 +1,7 @@
+namespace Laerdal.McuMgr.FileDownloading
+{
+    public partial class FileDownloader
+    {
+        public bool TryResume() => _nativeFileDownloaderProxy?.TryResume() ?? false;
+    }
+}
